[Consensus Observer] Move subscriptions to separate file.
