+++ consensus/src/consensus_observer/mod.rs
@@ -9,3 +9,4 @@ pub mod network_events;
 pub mod network_message;
 pub mod observer;
 pub mod publisher;
+mod subscription;
